Convert gmm implementations.
