+++ tools/adol-c/Makefile
@@ -4,15 +4,11 @@ CFLAGS=-std=c++17 -O3 -Wall -I../../cpp
 LDFLAGS=-lm -ladolc
 
 UTIL_OBJECTS=../../cpp/adbench/shared/utils.o ../../cpp/adbench/io.o
-GMM_OBJECTS=AdolCGMM.o run_gmm.o
 HT_OBJECTS=AdolCHT.o run_ht.o
 LSTM_OBJECTS=AdolCLSTM.o run_lstm.o
 EXECUTABLES=run_hello run_gmm run_ba run_ht run_lstm
 
 all: $(EXECUTABLES)
-
-run_gmm: $(GMM_OBJECTS) $(UTIL_OBJECTS)
-	$(CXX) -o $@ $^ $(LDFLAGS)
 
 run_ht: $(HT_OBJECTS) $(UTIL_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
@@ -27,4 +23,4 @@ run_%: run_%.cpp
 	$(CXX) $< -c -o $@ $(CFLAGS)
 
 clean:
-	rm -f $(UTIL_OBJECTS) $(GMM_OBJECTS) $(HT_OBJECTS) $(EXECUTABLES)
+	rm -f $(UTIL_OBJECTS) $(HT_OBJECTS) $(EXECUTABLES)
