+++ tools/enzyme/Makefile
@@ -7,17 +7,11 @@ ENZYME_LIB?=/home/gradbench/enzyme-build/Enzyme/
 LLDENZYME=$(ENZYME_LIB)/LLDEnzyme-19.so
 
 UTIL_OBJECTS=../../cpp/adbench/shared/utils.o ../../cpp/adbench/io.o
-GMM_OBJECTS=EnzymeGMM.o run_gmm.o
 LSTM_OBJECTS=EnzymeLSTM.o run_lstm.o
-BA_OBJECTS=run_ba.o
 HT_OBJECTS=EnzymeHT.o run_ht.o
-KMEANS_OBJECTS=run_kmeans.o
 EXECUTABLES=run_hello run_gmm run_lstm run_ba run_ht run_kmeans
 
 all: $(EXECUTABLES)
-
-run_gmm: $(GMM_OBJECTS) $(UTIL_OBJECTS)
-	$(CXX) -o $@ $^ $(LDFLAGS)
 
 run_lstm: $(LSTM_OBJECTS) $(UTIL_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
@@ -32,4 +26,4 @@ run_%: run_%.cpp
 	$(CXX) $< -c -o $@ $(CXXFLAGS) -I../../cpp
 
 clean:
-	rm -f $(UTIL_OBJECTS) $(HELLO_OBJECTS) $(GMM_OBJECTS) $(LSTM_OBJECTS) $(HT_OBJECTS) $(KMEANS_OBJECTS) $(EXECUTABLES)
+	rm -f $(UTIL_OBJECTS) $(LSTM_OBJECTS) $(HT_OBJECTS) $(EXECUTABLES)
