+++ tools/manual/Makefile
@@ -4,16 +4,12 @@ CFLAGS=-std=c++17 -O3 -Wall -I../../cpp
 LDFLAGS=-lm
 
 UTIL_OBJECTS=../../cpp/adbench/shared/utils.o ../../cpp/adbench/io.o
-GMM_OBJECTS=ManualGMM.o run_gmm.o gmm_d.o
 LSTM_OBJECTS=ManualLSTM.o run_lstm.o lstm_d.o
 HT_OBJECTS=ManualHT.o run_ht.o ht_d.o
 KMEANS_OBJECTS=run_kmeans.o
 EXECUTABLES=run_hello run_gmm run_ba run_lstm run_ht run_kmeans
 
 all: $(EXECUTABLES)
-
-run_gmm: $(GMM_OBJECTS) $(UTIL_OBJECTS)
-	$(CXX) -o $@ $^ $(LDFLAGS)
 
 run_lstm: $(LSTM_OBJECTS) $(UTIL_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
@@ -28,4 +24,4 @@ run_%: run_%.cpp
 	$(CXX) $< -c -o $@ $(CFLAGS)
 
 clean:
-	rm -f $(UTIL_OBJECTS) $(GMM_OBJECTS) $(LSTM_OBJECTS) $(HT_OBJECTS) $(KMEANS_OBJECTS) $(EXECUTABLES)
+	rm -f $(UTIL_OBJECTS) $(LSTM_OBJECTS) $(HT_OBJECTS) $(KMEANS_OBJECTS) $(EXECUTABLES)
