+++ tools/manual/gmm_d.hpp
@@ -1,0 +1,180 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+// Based on:
+//   https://github.com/microsoft/ADBench/blob/38cb7931303a830c3700ca36ba9520868327ac87/src/cpp/modules/manual/gmm_d.h
+//   https://github.com/microsoft/ADBench/blob/38cb7931303a830c3700ca36ba9520868327ac87/src/cpp/modules/manual/gmm_d.cpp
+
+#pragma once
+
+#include <cmath>
+#include <vector>
+#include <algorithm>
+
+#include "adbench/shared/defs.h"
+#include "adbench/shared/matrix.h"
+
+#include "gradbench/evals/gmm.hpp"
+
+// d: dim
+// k: number of gaussians
+// n: number of points
+// alphas: k logs of mixture weights (unnormalized), so
+//          weights = exp(log_alphas) / sum(exp(log_alphas))
+// means: d*k component means
+// icf: (d*(d+1)/2)*k inverse covariance factors 
+//                  every icf entry stores firstly log of diagonal and then 
+//          columnwise other entries
+//          To generate icf in MATLAB given covariance C :
+//              L = inv(chol(C, 'lower'));
+//              inv_cov_factor = [log(diag(L)); L(au_tril_indices(d, -1))]
+// wishart: wishart distribution parameters
+// x: d*n points
+// err: objective function output
+// J: gradient output
+void gmm_objective_d(int d, int k, int n,
+                     const double *alphas,
+                     const double *means,
+                     const double *icf,
+                     const double *x,
+                     gmm::Wishart wishart,
+                     double *err,
+                     double *J);
+
+void Qtransposetimesx(int d,
+                      const double* const Ldiag,
+                      const double* const icf,
+                      const double* const x,
+                      double* Ltransposex) {
+  int Lparamsidx = d;
+  for (int i = 0; i < d; i++)
+    Ltransposex[i] = Ldiag[i] * x[i];
+
+  for (int i = 0; i < d; i++)
+    for (int j = i + 1; j < d; j++)
+      {
+        Ltransposex[i] += icf[Lparamsidx] * x[j];
+        Lparamsidx++;
+      }
+}
+
+void compute_q_inner_term(int d,
+                          const double* const Ldiag,
+                          const double* const xcentered,
+                          const double* const Lxcentered,
+                          double* logLdiag_d) {
+  for (int i = 0; i < d; i++) {
+    logLdiag_d[i] = 1. - Ldiag[i] * xcentered[i] * Lxcentered[i];
+  }
+}
+
+void compute_L_inner_term(int d,
+                          const double* const xcentered,
+                          const double* const Lxcentered,
+                          double* L_d) {
+  int Lparamsidx = 0;
+  for (int i = 0; i < d; i++) {
+    int n_curr_elems = d - i - 1;
+    for (int j = 0; j < n_curr_elems; j++) {
+      L_d[Lparamsidx] = -xcentered[i] * Lxcentered[d - n_curr_elems + j];
+      Lparamsidx++;
+    }
+  }
+}
+
+double logsumexp_d(int n, const double* const x, double *logsumexp_partial_d)
+{
+  int max_elem = arr_max_idx(n, x);
+  double mx = x[max_elem];
+  double semx = 0.;
+  for (int i = 0; i < n; i++) {
+    logsumexp_partial_d[i] = exp(x[i] - mx);
+    semx += logsumexp_partial_d[i];
+  }
+  if (semx == 0.) {
+    std::fill(logsumexp_partial_d, logsumexp_partial_d + n, 0.0);
+  }
+  else {
+    logsumexp_partial_d[max_elem] -= semx;
+    for (int i = 0; i < n; i++)
+      logsumexp_partial_d[i] /= semx;
+  }
+  logsumexp_partial_d[max_elem] += 1.;
+  return log(semx) + mx;
+}
+
+void gmm_objective_d(int d, int k, int n,
+                     const double *alphas,
+                     const double *means,
+                     const double *icf,
+                     const double *x,
+                     gmm::Wishart wishart,
+                     double *err,
+                     double *J)
+{
+  const double CONSTANT = -n * d*0.5*log(2 * PI);
+  int icf_sz = d * (d + 1) / 2;
+
+  std::vector<double> Qdiags(d*k);
+  std::vector<double> sum_qs(k);
+  std::vector<double> main_term(k);
+  std::vector<double> xcentered(d);
+  std::vector<double> Qxcentered(d);
+
+  gmm::preprocess_qs(d, k, icf, sum_qs.data(), Qdiags.data());
+
+  std::fill(J, J + (k + d * k + icf_sz * k), 0.0);
+
+  std::vector<double> curr_means_d(d*k);
+  std::vector<double> curr_q_d(d*k);
+  std::vector<double> curr_L_d((icf_sz - d) * k);
+
+  double *alphas_d = J;
+  double *means_d = &J[k];
+  double *icf_d = &J[k + d * k];
+
+  double slse = 0.;
+  for (int ix = 0; ix < n; ix++) {
+    const double* const curr_x = &x[ix*d];
+    for (int ik = 0; ik < k; ik++) {
+      int icf_off = ik * icf_sz;
+      double *Qdiag = &Qdiags[ik*d];
+
+      subtract(d, curr_x, &means[ik*d], xcentered.data());
+      gmm::Qtimesx(d, Qdiag, &icf[ik*icf_sz + d], xcentered.data(), Qxcentered.data());
+      Qtransposetimesx(d, Qdiag, &icf[icf_off], Qxcentered.data(), &curr_means_d[ik*d]);
+      compute_q_inner_term(d, Qdiag, xcentered.data(), Qxcentered.data(), &curr_q_d[ik*d]);
+      compute_L_inner_term(d, xcentered.data(), Qxcentered.data(), &curr_L_d[ik*(icf_sz - d)]);
+      main_term[ik] = alphas[ik] + sum_qs[ik] - 0.5*sqnorm(d, Qxcentered.data());
+    }
+    slse += logsumexp_d(k, main_term.data(), main_term.data());
+    for (int ik = 0; ik < k; ik++) {
+      int means_off = ik * d;
+      int icf_off = ik * icf_sz;
+      alphas_d[ik] += main_term[ik];
+      for (int id = 0; id < d; id++){
+        means_d[means_off + id] += curr_means_d[means_off + id] * main_term[ik];
+        icf_d[icf_off + id] += curr_q_d[ik*d + id] * main_term[ik];
+      }
+      for (int i = d; i < icf_sz; i++){
+        icf_d[icf_off + i] += curr_L_d[ik*(icf_sz - d) + (i - d)] * main_term[ik];
+      }
+    }
+  }
+
+  std::vector<double> lse_alphas_d(k);
+  double lse_alphas = logsumexp_d(k, alphas, lse_alphas_d.data());
+  for (int ik = 0; ik < k; ik++) {
+    alphas_d[ik] -= n * lse_alphas_d[ik];
+    for (int id = 0; id < d; id++) {
+      icf_d[ik*icf_sz + id] += wishart.gamma*wishart.gamma * Qdiags[ik*d + id] * Qdiags[ik*d + id]
+        - wishart.m;
+    }
+    for (int i = d; i < icf_sz; i++) {
+      icf_d[ik*icf_sz + i] += wishart.gamma*wishart.gamma*icf[ik*icf_sz + i];
+    }
+  }
+
+  *err = CONSTANT + slse - n * lse_alphas;
+  *err += gmm::log_wishart_prior(d, k, wishart, sum_qs.data(), Qdiags.data(), icf);
+}
