+++ tools/tapenade/Makefile
@@ -5,7 +5,7 @@ CXXFLAGS=-std=c++17 -O3 -Wall
 LDFLAGS=-lm
 
 UTIL_OBJECTS=utils/adBuffer.o utils/adStack.o ../../cpp/adbench/shared/utils.o ../../cpp/adbench/io.o
-GMM_OBJECTS=TapenadeGMM.o gmm/gmm.o gmm/gmm_b.o run_gmm.o
+GMM_OBJECTS=gmm/gmm.o gmm/gmm_b.o run_gmm.o
 BA_OBJECTS=ba/ba.o ba/ba_b.o run_ba.o
 LSTM_OBJECTS=TapenadeLSTM.o lstm/lstm.o lstm/lstm_b.o run_lstm.o
 HT_OBJECTS=TapenadeHT.o hand/hand.o hand/hand_d.o run_ht.o
@@ -13,7 +13,7 @@ EXECUTABLES=run_gmm run_ba run_lstm run_ht
 
 all: $(EXECUTABLES)
 
-run_ba: $(BA_OBJECTS) $(UTIL_OBJECTS)
+run_ba: $(BA_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
 
 run_gmm: $(GMM_OBJECTS) $(UTIL_OBJECTS)
